fix(run-once): reject missing branch artifacts
